feat: v4
- Global CLI, you do not need to wrap active downloads with downloadSequence
- Auto increase parallel stream, maximize download speed
- Not reusing redirected URL by default - prevent token expires for long downloads
- Performance & stability improvements
- Remote CLI progress, show download progress in another process easily

BREAKING CHANGE:
- `partsURL` removed in favor of `partURLs`
- Not reusing redirected URL by default
- Different chunk size based on programType
- You can recall `.download()` and it will not throw an error - will return the same promise of downloading (do *not* redownload the file)
- You can change the parallel stream count after download has started

+++ test/daynamic-content-length.test.ts
@@ -15,10 +15,12 @@ describe("Dynamic content download", async () => {
     beforeAll(async () => {
         regularDownload = await ensureLocalFile(DYNAMIC_DOWNLOAD_FILE, ORIGINAL_FILE);
         originalFileHash = await fileHash(regularDownload);
-    });
+    }, 1000 * 30);
 
     afterAll(async () => {
-        await fs.remove(regularDownload);
+        if (regularDownload) {
+            await fs.remove(regularDownload);
+        }
     });
 
 
@@ -26,7 +28,11 @@ describe("Dynamic content download", async () => {
         const downloader = await downloadFile({
             url: DYNAMIC_DOWNLOAD_FILE,
             directory: ".",
-            fileName: IPUll_FILE
+            fileName: IPUll_FILE,
+            defaultFetchDownloadInfo: {
+                acceptRange: false,
+                length: 0
+            }
         });
 
         await downloader.download();
@@ -39,7 +45,11 @@ describe("Dynamic content download", async () => {
 
     test.concurrent("Browser Download", async (context) => {
         const downloader = await downloadFileBrowser({
-            url: DYNAMIC_DOWNLOAD_FILE
+            url: DYNAMIC_DOWNLOAD_FILE,
+            defaultFetchDownloadInfo: {
+                acceptRange: false,
+                length: 0
+            }
         });
 
         await downloader.download();
